feat: export publishInterchainToken function in tx-builder
